bug: go1.17 support
Fuzzing tests are introduced in go1.18. When linter runs on older go runtime, it's ok, that testing.F is not found.

Disable f-checks if testing.F not found.
